requirements.txt

discord.py>=2.3.0
pytz
pytz
pytz
pytz
